putObject: make `size` param optional (#494)
This commit allows users to put streams without knowing the size of the
stream beforehand, a use case very common in nodejs APIs.

Under the hood, this commit adds a new way of uploading, the
ObjectUploader. This class handles uploads in their entirety: it handles
hashes, resuming and finishing uploads, executing the callback,
everything, while inheriting from a stream. This allows for true
streaming, as opposed to faked streaming as before.

This includes a breaking change:
- With the old object uploader, if we were told to upload a stream but the
size given was larger than the stream size, we would leave the multipart
upload open on the server.
- Now, the object uploader will automatically finish the old multipart
upload if the size was larger than the stream size.

This also removes content-type from the documentation because it really
had/has no point.