Allow bucket region cache to be overriden.
Add region= option to our API constructor. This option
is added to avoid making calls to server when an
application written using `minio-js` is only meant
to generate presigned URLs which is a quite common use
case.

Setting this option - overrides any future region
changes and makes the API more stricter. For example
if you are going to use your initialized client
against a bucket which exists in a different region
all API operations would fail.