ci: enable unit and functional tests
This commit re-enables both the unit tests and functional tests for the
CI machine, so it'll run automatically for PRs and commits. This will
ensure commits don't accidentally break tests and have unintended side
effects.

Functional tests are run against the play.minio.io server.

This commit also fixes issues in the tests that would have caused them
to fail. It also allows functional tests to be run on non-mint machines
and the CI after b3b05e0, which introduced a few bugs.