add promise support for API
This commit adds promise support for the entire JS API. It will continue
to work as expected with a callback when one is passed; when one is not,
it will return a Promise. These functions would always error when a
callback was not called so this should be entirely backwards compatible.