change linter to eslint, fix codebase (#606)
* lint: add eslint, remove jsc and jshint

JSHint as it was wasn't strict at all. JSC, on the other hand, was
discontinued and continued into eslint, a very popular tool. I chose to
replace them with eslint for its strictness. The tools are pretty
compatible with eachother so there wasn't much of a choice to be made,
they'd all work to some extent.

* lint: fix code style to match eslint rules

mostly semicolons to remove, unused variable declarations

* lint: fix indentation