Fix Hashing bug and remove putobject string and list 1001 objects tests
Removing putobject string buffer test as it is incorrect.
There are chances that some unicode character can come into
play and the test might occasionally fail.

Fixed a Hash calculation bug in getpartialobject test, seen in Mint.
The code assumed that the first 100KB of data in the 6MB object will
have the same size as a 100KB object itself.

The listobjects test to list 1001 objects fails intermittently in
removeObject part of the test in Appveyor.
This will be tracked in a seperate issue.

Fixes #627 and #629