Skip content-MD5 for secure putobject call, add prefix to bucket in functional test (#631)
Not calulating Content-MD5 for secure, non-anonymous put object requests.

Added a prefix to the bucket name created by the functional test, so that
Mint can clean it up easily, if it has to.

Modifying appveyor.yml to not do "npm test", as "npm install" runs functional
and unit tests already.

Fixes #589, #632, #634