Ensure that tests are run on Appveyor
Due to 'npm prepublish' being deprecated, the relevant command was moved
to 'npm prepublishOnly', which doesn't run on local 'npm install' and
resulted in our tests not being ran on Appveyor. The functionality of
'npm prepublish' is preserved on 'npm prepare', but that is already
reserved for users who install the package directly from the GitHub URL.
Therefore, an additional line has been added to appveyor.yml that will
run the tests.