Ensure returned versionId is a string when listing objects
Returned versionId is a string when calling statObject(), as it is read
without transformation (casting) from the headers.

However, the xml parser casts automatically what looks like a number in
the responses. Consequentially, when the S3 storage provider returns a
VersionId with only numbers, listObjects() returns an object with a
versionId as a number, which is inconsistent with what returns other
methods like statObjects() and may result in erratic behaviors for the
program using these methods.